requirements.txt
streamlit>=1.35.0
pandas>=2.2.0
plotly>=5.24.0
python-dotenv>=1.0.1
requests>=2.32.0
